Beginnings of the Cron Parser
I think I'm going to end up renaming some stuff once I get a better
sense of how it's going to work - in particular, I'm probably going
to try and dissuade the notion that this is a general purpose crontab
parsing library, rather than something that is fairly specialized.
For now though, things are moving along. I'm using this as an excuse
to finally try out the doctest module, which I've never gotten around
to using before. Also TDD.