Added Job.next_value + refactoring
Added a couple of fields to Job, and I added carry flags to next_value.
And the big change was adding Job.next_value. The two biggest goals for
this code were avoiding code duplication, especially of ugly
special-cased stuff, and offloading all date calculations to python's
datetime class. This means that I use the crontab values to generate
candidate dates, then try and actually create a datetime instance. If
the date doesn't exist, then this will fail and we need to move onto
the next possible date. I'm fairly pleased with how it works, if not how
it looks.