Major refactoring of Job.next_value
I gutted it because I discovered that it didn't work if the month was
held constant but the year wasn't... Sigh. This is one of the reasons I
find working with dates to be a pain. On the bright side, it is actually
almost readable now and I'm fairly happy with the refactor. On the down
side again, it actually isn't done. It doesn't handle day of week
specifications at all - but all my tests pass, which means that my tests
were missing this case before. Also a slight tweak to Range.validate.